by default index uri as a keyword field
it makes little sense for the uri to be indexed as a free text field, this seems
to be a simple ommission when the uri field was added. I doubt it will break
existing searches but this should be properly tested. In most cases I expect it
to improve search performance if the uri field is used for matching